% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeans.R
\name{bloodmeans}
\alias{bloodmeans}
\title{Extract Mean Values of Blood Biomarkers}
\usage{
bloodmeans(df, var = NULL)
}
\arguments{
\item{df}{a \code{data.frame} with at least one numeric variable in order to
get a non-empty result.}

\item{var}{NULL (default) or the unquoted name of the variable by which to
group the input df.}
}
\value{
a data.frame with the mean values of each numeric
}
\description{
This function accepts a dataframe as input and extracts the mean value of
each numeric variable.
}
\examples{
\dontrun{
library(magrittr)
data.frame(x1 = c(1,2,3), x2 = c(4,5,6)) \%>\%
  bloodstats::bloodmeans()

# Grouped bloodmeans
bloodstats::df_example \%>\%
  bloodstats::bloodmeans(var = group)
}
}
\author{
John Doe
}
