% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_example}
\alias{df_example}
\title{Random Metabolic Data}
\format{A data frame (tibble) with 100 rows and 4 columns.
\describe{
  \item{id}{ID of individual.}
  \item{biofluid}{Biofluid type, blood or urine}
  \item{males}{Log odds for incident type 2 diabetes.}
  \item{females}{Standard error.}
}}
\source{
Generated with a random \code{set.seed(34)}
}
\usage{
df_example
}
\description{
A dataframe with random data for blood and urine, males and females.
}
\keyword{datasets}
