% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeans.R
\name{bloodmeans}
\alias{bloodmeans}
\title{Extract Mean Values of Blood Biomarkers}
\usage{
bloodmeans(df)
}
\arguments{
\item{df}{a \code{data.frame} with at least one numeric variable in order to
get a non-empty result.}
}
\value{
a data.frame with the mean values of each numeric
}
\description{
This function accepts a dataframe as input and extracts the mean value of
each numeric variable.
}
\examples{
library(magrittr)
data.frame(x1 = c(1,2,3), x2 = c(4,5,6)) \%>\%
  bloodstats::bloodmeans()
}
\author{
John Doe
}
